✨ Get theme color
