+++ src/frontend/ui/common/index.ts
@@ -1,6 +1,7 @@
 import Button from './button'
 import Checkbox from './checkbox'
+import Favicon from './favicon'
 import { Ring } from './ring'
 import { Switch } from './switch'
 
-export { Button, Checkbox, Ring, Switch }
+export { Button, Checkbox, Ring, Switch, Favicon }
